Fix G35 on Arduino
The tramming_point_name[] array is PROGMEM and points to PGMSTR()
strings... this apparently means it can't be used with
SERIAL_ECHOPAIR() on Arduino.

On my CR-10 v3 with original Creality board, the first output sent
a long string of characters to the serial port and then hung.

This change allows it to work correctly.

NOTE: The debug output wasn't tested, and I'm sure there's a better
way of doing this, but I'm not familiar enough with the firmware to
know what SERIAL_*() macro is appropriate for this.